[LIVY-406][SERVER] Fix Livy cannot app id in yarn client mode issue
In our `SparkYarnApp` logic, we have to wait for process to exit before to query app id. Since the process is never exited in yarn client mode, so it will block follow-up logic to get app id and other yarn application information.

So here propose to remove this logic, also because now we will query app id ever since app is launched, so we should increase the look-up time to avoid timeout.

Author: jerryshao <sshao@hortonworks.com>

Closes #50 from jerryshao/LIVY-406.