[LIVY-406][LIVY-406][FOLLOWUP][SERVER] Fix UT failure in branch 0.4 and apply to Master branch for consistency
Because we changed the AppState logic in #39 , so the UT which passed in master branch will be failed in branch 0.4, here to fix this issue.

Also apply this PR to master branch for the consistency, it should also be worked in master branch.

Author: jerryshao <sshao@hortonworks.com>

Closes #52 from jerryshao/LIVY-406-followup.

(cherry picked from commit 0ca16d5d2fb9bed6c66a6a787afb5a7f025c577f)
Signed-off-by: jerryshao <sshao@hortonworks.com>