[LIVY-7] added autocompletion api and implementation for scala
I started an implementation of the very old feature request (LIVY-7) for code autocompletion.
This implementation works with scala 2.11 and scala 2.10.

I'd be happy if somebody could review and comment it.

As for the API: I chose a synchronous call because resolving the code options shouldn't be a very long process (and if it were it wouldn't make sense anyway).

Author: Pascal Pellmont <github@ppo2.ch>

Closes #51 from pellmont/master.