[LIVY-415] Use objects and abstract classes in for Kind and SessionState.
## What changes were proposed in this pull request?

- Use a singleton (object) rather than having a case class and creating an object of only one type whenever we need an object of the type. Using an object helps with code efficiency and readability.
- Use abstract classes to set up methods with default definitions to follow the DRY principles.

## How was this patch tested?

All the existing tests were modified to use the new changes. The changes were then tested on Travis with existing settings to make sure it's compatible with the the upstream repository.

Please review https://livy.incubator.apache.org/community/ before opening a pull request.

I have read through the page and felt like this change does not require a JIRA. If you feel otherwise, I can create a JIRA and then attach it to the PR.

Author: Arun Allamsetty <arun.allamsetty@gmail.com>

Closes #62 from aa8y/singleton.