[LIVY-415][CORE][FOLLOWUP] Fix static object creation issue for SessionState
## What changes were proposed in this pull request?

In LIVY-415, we changed `SessionState` to static object instead of case class. This is OK for `SessionState`, but for `FinishedSessionState`, it requires `time` field to be set during object creation. Using static object will never reflect the actual `time` when object is created. So here propose to fix it.

CC aa8y please take a review, this is introduced in your changes.

## How was this patch tested?

Existing tests.

Author: jerryshao <sshao@hortonworks.com>

Closes #67 from jerryshao/LIVY-415-followup.