[LIVY-472][SERVER] Improve the logs for fail-to-create session
## What changes were proposed in this pull request?

Livy currently doesn't give a very clear log about the fail-to-create session, it only says that session related app tag cannot be found in RM, but doesn't tell user how to search and get the true root cause. So here change the logs to make it more clear.

## How was this patch tested?

Local verification.

Author: jerryshao <sshao@hortonworks.com>

Closes #96 from jerryshao/LIVY-472.