Fix incorrect default value in livy-client.conf.template
The template for `livy-client.conf` implies that the default value of `launcher.port.range` is `10000~10110`, but this is in fact incorrect - the real default value is `10000~10010`, as you can [see here](https://github.com/apache/incubator-livy/blob/master/rsc/src/main/java/org/apache/livy/rsc/RSCConf.java#L54). This is really confusing, especially since:
a) 10110 is wrong on the Spark side, the default configuration there won't open that many ports
b) 10010 and 10110 look very similar to the human eye so it can take a long time to notice the discrepancy

This commit modifies the default conf template to reflect the real default value.