[LIVY-543] Clean thriftserver dependencies from Hive's jetty
## What changes were proposed in this pull request?

Hive 3.0 uses Jetty 9.3.8. So there are some jars coming from hive dependency which are conflicting with our Jetty dependencies which were upgraded in LIVY-526. Hence we should ignore the dependencies Hive has on Jetty and use our own version. After the refactor which has been done in the thriftserver, this should be easy. This can cause issue when running the server (not always, though, which makes it hard to debug and reproduce).

The PR removes also an extra-dependency which was left in the thriftserver-server module which is not needed anymore after the refactor which created the thriftserver-session module.

## How was this patch tested?

manual tests, checking the generated JARs

Please review https://livy.incubator.apache.org/community/ before opening a pull request.

Author: Marco Gaido <mgaido@hortonworks.com>

Closes #134 from mgaido91/LIVY-543.