[LIVY-551] Add "doAs" impersonation support
## What changes were proposed in this pull request?

Currently `proxyuser` is used in batches and sessions for impersonation. This should be extended to support impersonation for all endpoints. Adding the query parameter `doAs` matches others in the Hadoop ecosystem and will work across endpoints.

https://issues.apache.org/jira/browse/LIVY-551

## How was this patch tested?

Added unit and integration tests for this feature. Manually tested to ensure that `proxyuser` still works. Also checked that if both `proxyuser` and the new `doAs` parameter are specified that `doAs` takes precedence.

Author: Kevin Risden <krisden@apache.org>

Closes #141 from risdenk/LIVY-551.