[LIVY-591] Changed ACLs enforcement to occur on proxy user, if available
ACLs enforcement occurs by checking whether the user requesting for a
resource/access is owner of that resource or has sufficient permissions.

This check used to occur against the session owner, even when a proxy user was
specified during the creation of the resource. This can cause a issue where a
proxy user can create a resource but can't access or modify it.

The check has been modified to occur againt the session proxy user, defaulting
to session owner if proxy user is not specified.

Testing Done:
Added a couple of unit tests