[LIVY-582][TESTS] Hostname in python-api test should be lower case to avoid test failures
## What changes were proposed in this pull request?
In the python-API test code, when returning a mocked response, the mock lib will compare the URL with the predefined URLs case sensitive. However, the `Request` lib used in the Livy python API will change the URL to lower case. This will cause test failures on a machine with an upper case hostname.

This patch turns the hostname in python-API test code into the lower case to avoid such test failures.

## How was this patch tested?
Existing test. Run test specifically on a machine with an upper case hostname.

Author: Yiheng Wang <yihengw@gmail.com>

Closes #180 from yiheng/fix_582.