[LIVY-571] cleanupStatement should not throw exception when statementId not exist
## What changes were proposed in this pull request?

`ThriftSessionState` is used to store query result by statementId. When an exception is thrown during execute query, no query result is stored. But when a statement is closed from beeline, a request is invoked to remove cached query result in `ThriftSessionState`.

Remove statement query result from `ThriftSessionState` currently throws an exception, hiding the original query failure exception and message. The PR makes it return cleanly, so the proper exception is reported to the end user.

## How was this patch tested?

Added unit tests.

Author: Jeffrey(Xilang) Yan <7855100+yantzu@users.noreply.github.com>
Author: yantzu <7855100+yantzu@users.noreply.github.com>
Author: 80254702 <80254702@adc.com>

Closes #182 from yantzu/LIVY-571.