LIVY-615: Validate livy.ui.basePath property
The default value of livy.ui.basePath has been changed from empty string
to "/" instead. As the property was not working before this should be
a safe change. Nonetheless, if empty string is used for this property
things will still work as expected. Any value that does not start with
a slash will fail however.

Changes have also been made to make things pass the scala style check.