[LIVY-617] Livy session leak on Yarn when creating session duplicated names
## What changes were proposed in this pull request?
When creating a session with duplicated name, instead of throw exception in SessionManager.register() method, we should stop the session. Otherwise the session driver process will keep running and end up creating a leaked Yarn application.

https://issues.apache.org/jira/browse/LIVY-617

## How was this patch tested?

This is just a simple fix and verified with manual end to end test.

Author: Shanyu Zhao <shzhao@microsoft.com>

Closes #187 from shanyu/shanyu.