[LIVY-633][SERVER] session should not be gc-ed for long running queries
## What changes were proposed in this pull request?
Currently, Livy records the last activity time of the session before statement execution. If a statement runs too long, exceeding then the session timeout, the session will be garbage collected after the statement execution.

This should not be the expected behavior. The statement execution time should not be count into idle. We should update the last activity time after the statement execution.

We cannot be updated when session changes state from busy to idle in the Session class. So in this patch, we add a replLastActivity field into the rscClient, which will be updated when the repl state changes. So when session changes its state from busy to idle, this field will catch the time and finally reflect on the session last activity.

## How was this patch tested?
Manual test. Also, add a new unit test.

Existing unit tests and integration tests.

Author: yihengwang <yihengwang@tencent.com>
Author: Yiheng Wang <yihengw@gmail.com>

Closes #224 from yiheng/fix_633.