[LIVY-717] introduce maven property to set ZooKeeper version
When we want to use Livy in a cluster where a newer ZooKeeper server version is used, we might run into run-time errors if we compile Livy using the current Curator / Hadoop versions. The Curator version can already explicitly set with the `curator.version` maven property in build time, but we were still missed the same parameter for ZooKeeper.

In this PR I added a new maven parameter called `zookeeper.version` and after analyzing the maven dependency tree, I made sure that the Curator and ZooKeeper versions used in compile time are always harmonized and controlled by the maven parameters.

I set the `zookeeper.version` in maven to `3.4.6` to be backward compatible with the current Livy dependencies.

see https://issues.apache.org/jira/browse/LIVY-717

Author: Mate Szalay-Beko <szalay.beko.mate@gmail.com>

Closes #262 from symat/LIVY-171.