[LIVY-745] Ensure that a single LivyClientFactory gets loaded.
In LivyClientBuilder, a ServiceLoader is used to load configured
LivyClientFactory providers (there is only one, RSCClientFactory).
Correct behavior of RSCClientFactory implicitly depends on there being
a single instance of it in the Livy server.
Instead of instantiating a new ServiceLoader every time the build method
is called on a LivyClientBuilder, keep a single ServiceLoader in a static
field. The ServiceLoader will lazily load the LivyClientFactory
(RSCClientFactory) and cache it. ServiceLoader instances are not safe
for use by multiple concurrent threads, so call the ServiceLoader's
iterator in a synchronized block.