[LIVY-748] Add support for running Livy Integration tests against secure external clusters
## What changes were proposed in this pull request?

https://issues.apache.org/jira/projects/LIVY/issues/LIVY-748?filter=reportedbyme

This PR configures the integration tests to work on Livy Servers instances that require basic or kerberos authentication. To do this we had to replace the async httpclient with apaches normal httpclient, as there were issues found getting kerberos authentication to work with the async client.

This PR also contains a minor fix to the InteractiveIT test suite. One of the tests would occasionally fail as a statement submitted to the interactive session would not be finished processing when the result was checked. The correctness check was also changed to correctly check for code completition

## How was this patch tested?
This patch was tested on local deployments of livy server, both on a instance that required kerberos authentication and on one that required basic authentication.

Author: Roger Liu <roliu@microsoft.com>
Author: roliu <roliu@microsoft.com>

Closes #278 from RogPodge/LivyKerberosIT.