[LIVY-754][THRIFT] Encode precision and scale for decimal type.
## What changes were proposed in this pull request?

When a `org.apache.livy.thriftserver.session.DataType.DECIMAL` is converted to a `org.apache.hive.service.rpc.thrift.TTypeDesc` for sending a Thrift response to a client request for result set metadata, the `TTypeDesc` contains a `TPrimitiveTypeEntry(TTypeId.DECIMAL_TYPE)` without `TTypeQualifiers` (which are needed to capture the precision and scale).
With this change, we include the qualifiers in the `TPrimitiveTypeEntry`. We use both the name and the `DataType` of a field type to construct the `TTypeDesc`. We are able to do this without changing the existing internal representation for data types because we can obtain the precision and scale from the name of the decimal type.

## How was this patch tested?

Use beeline to connect to the Thrift server. Do a select from a table with a column of decimal type.
Also extended an existing integration test.

Author: Wing Yew Poon <wypoon@cloudera.com>

Closes #288 from wypoon/LIVY-754.