[LIVY-756] Add Spark 3.0 and Scala 2.12 support
## What changes were proposed in this pull request?

This PR is based tprelle 's PR #289 , and address all the left issues in that PR:

1. multi-scala version support in one build (Scala 2.11 and 2.12 support).
2. make SparkR work.

Also reverts most of the unnecessary changes. Besides this PR remove the build below 2.4 (2.2, 2.3), since Spark 2.2 and 2.3 only ships with Scala 2.11, hard to maintain multiple version. But user could still use 2.2 and 2.3 without changes.

All credits to tprelle.

## How was this patch tested?

Run UT and IT with Spark 2.4.5 and 3.0.0 locally.

Author: jerryshao <jerryshao@tencent.com>

Closes #300 from jerryshao/LIVY-756.